requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-allure-adaptor>=1.7.10
allure-pytest>=2.13.2
jsonschema>=4.21.1
python-dotenv>=1.0.0
urllib3>=2.0.0
